+++ resources/views/components/menu-layout.blade.php
@@ -1,0 +1,15 @@
+@aware(['title'])
+
+    <nav>
+
+        <ul {{ $attributes }}>
+
+            @isset($title)
+                <li class="menu-title">{{ $title }}</li>
+            @endisset
+
+            {{ $slot }}
+
+        </ul>
+
+    </nav>